Eliminate race between 'select' and 'close(chan)'.
You learn something new everyday. Apparently, the 'select' statement
does not execute code in the 'case' block if the case's channel was
closed *before* the select statement was executed. This created a
race between starting and stopping the watcher.

As a result, the watcher loop would never exit if it was started
and immediately stopped.

The cleanest solution to this involves creating a temporary channel
(named 'ready') that blocks until the select statement is executed.
Since the select statement allows us to receive and send, we can
send to 'ready', and unblock the start caller. This ensures that
further stop calls come after the select statement has come into
scope, thus eliminating the race condition.